bacquest-backend/requirements.txt
fastapi[standard]
uvicorn
pymupdf
openai
google-generativeai
groq
python-dotenv
sqlalchemy
pydantic-settings
python-multipart
